Move `get_order`, `normalize_chunks`, and `MatrixArray` to `core.utils`
