+++ graphblas/core/ss/matrix.py
@@ -1,6 +1,5 @@
 import itertools
 import warnings
-from numbers import Integral, Number
 
 import numba
 import numpy as np
@@ -13,21 +12,23 @@ from ... import monoid
 from ...dtypes import _INDEX, BOOL, INT64, _string_to_dtype, lookup_dtype
 from ...exceptions import _error_code_lookup, check_status, check_status_carg
 from .. import NULL, ffi, lib
-from ..base import call, record_raw
+from ..base import call
 from ..scalar import Scalar, _as_scalar, _scalar_index
 from ..utils import (
     _CArray,
+    _MatrixArray,
     _Pointer,
+    get_order,
     get_shape,
     ints_to_numpy_buffer,
     libget,
+    normalize_chunks,
     output_type,
     values_to_numpy_buffer,
     wrapdoc,
 )
 from .config import BaseConfig
 from .descriptor import get_compression_descriptor, get_nthreads_descriptor
-from .utils import get_order
 
 ffi_new = ffi.new
 
@@ -177,96 +178,6 @@ def head(matrix, n=10, dtype=None, *, sort=False):
     return rows, cols, vals
 
 
-def normalize_chunks(chunks, shape):
-    """Normalize chunks argument for use by `Matrix.ss.split`.
-
-    Examples
-    --------
-    >>> shape = (10, 20)
-    >>> normalize_chunks(10, shape)
-    [(10,), (10, 10)]
-    >>> normalize_chunks((10, 10), shape)
-    [(10,), (10, 10)]
-    >>> normalize_chunks([None, (5, 15)], shape)
-    [(10,), (5, 15)]
-    >>> normalize_chunks((5, (5, None)), shape)
-    [(5, 5), (5, 15)]
-    """
-    if isinstance(chunks, (list, tuple)):
-        pass
-    elif isinstance(chunks, Number):
-        chunks = (chunks,) * len(shape)
-    elif isinstance(chunks, np.ndarray):
-        chunks = chunks.tolist()
-    else:
-        raise TypeError(
-            f"chunks argument must be a list, tuple, or numpy array; got: {type(chunks)}"
-        )
-    if len(chunks) != len(shape):
-        typ = "Vector" if len(shape) == 1 else "Matrix"
-        raise ValueError(
-            f"chunks argument must be of length {len(shape)} (one for each dimension of a {typ})"
-        )
-    chunksizes = []
-    for size, chunk in zip(shape, chunks):
-        if chunk is None:
-            cur_chunks = [size]
-        elif isinstance(chunk, Integral) or isinstance(chunk, float) and chunk.is_integer():
-            chunk = int(chunk)
-            if chunk < 0:
-                raise ValueError(f"Chunksize must be greater than 0; got: {chunk}")
-            div, mod = divmod(size, chunk)
-            cur_chunks = [chunk] * div
-            if mod:
-                cur_chunks.append(mod)
-        elif isinstance(chunk, (list, tuple)):
-            cur_chunks = []
-            none_index = None
-            for c in chunk:
-                if isinstance(c, Integral) or isinstance(c, float) and c.is_integer():
-                    c = int(c)
-                    if c < 0:
-                        raise ValueError(f"Chunksize must be greater than 0; got: {c}")
-                elif c is None:
-                    if none_index is not None:
-                        raise TypeError(
-                            'None value in chunks for "the rest" can only appear once per dimension'
-                        )
-                    none_index = len(cur_chunks)
-                    c = 0
-                else:
-                    raise TypeError(
-                        "Bad type for element in chunks; expected int or None, but got: "
-                        f"{type(chunks)}"
-                    )
-                cur_chunks.append(c)
-            if none_index is not None:
-                fill = size - sum(cur_chunks)
-                if fill < 0:
-                    raise ValueError(
-                        "Chunks are too large; None value in chunks would need to be negative "
-                        "to match size of input"
-                    )
-                cur_chunks[none_index] = fill
-        elif isinstance(chunk, np.ndarray):
-            if not np.issubdtype(chunk.dtype, np.integer):
-                raise TypeError(f"numpy array for chunks must be integer dtype; got {chunk.dtype}")
-            if chunk.ndim != 1:
-                raise TypeError(
-                    f"numpy array for chunks must be 1-dimension; got ndim={chunk.ndim}"
-                )
-            if (chunk < 0).any():
-                raise ValueError(f"Chunksize must be greater than 0; got: {chunk[chunk < 0]}")
-            cur_chunks = chunk.tolist()
-        else:
-            raise TypeError(
-                "Chunks for a dimension must be an integer, a list or tuple of integers, or None."
-                f"  Got: {type(chunk)}"
-            )
-        chunksizes.append(cur_chunks)
-    return chunksizes
-
-
 def _concat_mn(tiles, *, is_matrix=None):
     """Argument checking for `Matrix.ss.concat` and returns number of tiles in each dimension"""
     from ..matrix import Matrix, TransposedMatrix
@@ -324,16 +235,6 @@ def _concat_mn(tiles, *, is_matrix=None):
 
 def _as_matrix(x):
     return x._as_matrix() if hasattr(x, "_as_matrix") else x
-
-
-class MatrixArray:
-    __slots__ = "_carg", "_exc_arg", "name"
-
-    def __init__(self, matrices, exc_arg=None, *, name):
-        self._carg = matrices
-        self._exc_arg = exc_arg
-        self.name = name
-        record_raw(f"GrB_Matrix {name}[{len(matrices)}];")
 
 
 class MatrixConfig(BaseConfig):
@@ -539,7 +440,7 @@ class ss:
         call(
             "GxB_Matrix_split",
             [
-                MatrixArray(tiles, self._parent, name="tiles"),
+                _MatrixArray(tiles, self._parent, name="tiles"),
                 _as_scalar(m, _INDEX, is_cscalar=True),
                 _as_scalar(n, _INDEX, is_cscalar=True),
                 _CArray(tile_nrows),
@@ -580,7 +481,7 @@ class ss:
             "GxB_Matrix_concat",
             [
                 self._parent,
-                MatrixArray(ctiles, name="tiles"),
+                _MatrixArray(ctiles, name="tiles"),
                 _as_scalar(m, _INDEX, is_cscalar=True),
                 _as_scalar(n, _INDEX, is_cscalar=True),
                 None,
